Checking if persons and tax are busy

+++ TaxInspection/Windows/AllTaxes.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using System.Windows;
-using Finisar.SQLite;
 using TaxInspection.Database_elements;
 
 namespace TaxInspection.Windows
@@ -24,11 +23,39 @@ namespace TaxInspection.Windows
 
             if (item != null)
             {
+                if (checkIfTaxIsBusy(item))
+                {
+                    MessageBox.Show("Податок вже існує в базі!");
+                    return;
+                }
+
                 Extensions.Tools.ExecuteQuery("DELETE FROM Taxes WHERE Id = " + item.TaxId);
 
                 ((App)Application.Current).Taxes.Remove(item);
             }
         }
+
+        private bool checkIfTaxIsBusy(Tax tax)
+        {
+            for(int i = 0; i < ((App)Application.Current).TaxesPayedByNatPersons.Count; i++)
+            {
+                var item = ((App)Application.Current).TaxesPayedByNatPersons[i];
+
+                if (item.TaxId == tax.TaxId)
+                    return true;
+            }
+
+            for (int i = 0; i < ((App)Application.Current).TaxesPayedByJurPersons.Count; i++)
+            {
+                var item = ((App)Application.Current).TaxesPayedByJurPersons[i];
+
+                if (item.TaxId == tax.TaxId)
+                    return true;
+            }
+
+            return false;
+        }
+
     }
 
 }
