+++ TaxInspection/Windows/NaturalPersons.xaml.cs
@@ -17,6 +17,12 @@ namespace TaxInspection.Windows
             var item = ListOfNaturalPersons.SelectedItem as NaturalPerson;
             if (item != null)
             {
+                if(checkIfPersonIsBusy(item))
+                {
+                    MessageBox.Show("Ця особа вже записана в базі сплачених податків!");
+                    return;
+                }
+
                 Extensions.Tools.ExecuteQuery("DELETE FROM NaturalPersons WHERE Id = " + item.Id);
 
                 ((App)Application.Current).NaturalPersons.Remove(item);
@@ -28,5 +34,19 @@ namespace TaxInspection.Windows
             CreateNewNaturalPerson window = new CreateNewNaturalPerson();
             window.Show();
         }
+
+        private bool checkIfPersonIsBusy(NaturalPerson person)
+        {
+            for (int i = 0; i < ((App)Application.Current).NaturalPersons.Count; i++)
+            {
+                var item = ((App)Application.Current).TaxesPayedByNatPersons[i];
+
+                if (item.PayerId == person.Id)
+                    return true;
+            }
+
+            return false;
+        }
+
     }
 }
